+++ packages/authentication/modules/access/useLogout/index.ts
@@ -1,4 +1,4 @@
-import { COOKIE_NAME } from '@baseapp-frontend/utils'
+import { ACCESS_COOKIE_NAME, REFRESH_COOKIE_NAME } from '@baseapp-frontend/utils'
 
 import { useQueryClient } from '@tanstack/react-query'
 import Cookies from 'js-cookie'
@@ -7,11 +7,16 @@ import { MFA_API_KEY } from '../../../services/mfa'
 import { USER_API_KEY } from '../../../services/user'
 import { ILogoutOptions } from './types'
 
-const useLogout = ({ cookieName = COOKIE_NAME, onLogout }: ILogoutOptions = {}) => {
+const useLogout = ({
+  cookieName = ACCESS_COOKIE_NAME,
+  refreshCookieName = REFRESH_COOKIE_NAME,
+  onLogout,
+}: ILogoutOptions = {}) => {
   const queryClient = useQueryClient()
 
   const logout = () => {
     Cookies.remove(cookieName)
+    Cookies.remove(refreshCookieName)
     queryClient.resetQueries(USER_API_KEY.getUser())
     queryClient.resetQueries(MFA_API_KEY.default)
     onLogout?.()